feat: tariff zones in map
